requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
SQLAlchemy==2.0.32
openpyxl==3.1.5
Flask-WTF==1.2.1
python-docx==1.1.2
